move pairwise_evaluation sample and test to a dedicated directory
